Revert "Fix alarm not firing in memory-pressure situations"
I have two different devices, both of which are experiencing issues
with randomly missed alarms in N. I can see in the logs that the
device wakes up in time, and it's aware that it has to fire an
alarm. So it gives the order to fire, but the order never reaches the
firing squad.

Testing this revert, both by myself and users, shows that not a single
alarm is missed. It wakes up and fires the alarm as it should every
time.

Note: CAF hijacked some of the code that Google removed in the patch,
and modified it for their own use, so I had to slightly adapt the
revert to that.

This reverts commit b76aa50f9e13e71e81b9d02cc3ebebaa73188d3e.

Change-Id: I3eeb85dec208cd0022c3d037a403bbe1a5d86e28